Always extract all files from zip when updating (related to issue #196)

+++ Views/Stats.cs
@@ -2354,7 +2354,9 @@ namespace FallGuysStats {
                                         if (entry.Name.IndexOf(".exe", StringComparison.OrdinalIgnoreCase) > 0) {
                                             exeName = entry.Name;
                                         }
-                                        File.Move(entry.Name, $"{entry.Name}.bak");
+                                        if (File.Exists(entry.Name)) {
+                                            File.Move(entry.Name, $"{entry.Name}.bak");
+                                        }
                                         entry.ExtractToFile(entry.Name, true);
                                     }
                                 }
